separated modal code structure
